feat: set service_name from Log Group and hostname from Log Stream
This sets the New Relic Logs attribute `service_name` to the CloudWatch
Log Group (trailing portion `/` delimited), and sets the attribute
`hostname` to the trailing portion of the CloudWatch Log Stream (`/` or
`]` delimited, the ECS task ID for ECS logs or identifier for an
instance of a Lambda).  Without this change, both of those fields are
N/A in New Relic Logs if the application is not setting them in the log
message.